pyproject.toml
[project]
name = "stock_picker"
version = "0.1.0"
description = "AI-powered stock analysis platform that identifies trending companies, conducts financial research, and provides investment recommendations"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.152.0,<1.0.0",
    "streamlit>=1.38.0,<2.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "pysqlite3-binary>=0.5.3"
]

[project.scripts]
stock_picker = "stock_picker.main:run"
run_crew = "stock_picker.main:run"
train = "stock_picker.main:train"
replay = "stock_picker.main:replay"
test = "stock_picker.main:test"
ui = "stock_picker.ui.app:main"
streamlit_app = "stock_picker.ui.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
